backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
apscheduler==3.10.4
aiosqlite==0.20.0
PyJWT==2.10.1
bcrypt==4.2.1
argon2-cffi==23.1.0
cryptography==44.0.0
slowapi==0.1.9
pandas==2.2.3
numpy==2.2.1
scipy==1.15.0
python-multipart==0.0.20
yfinance==0.2.51
